migration/requirements.txt
mysql-connector-python
clickhouse-driver
tqdm
